add migrations, timestamps

+++ db/sqlite/queryBuilder/delete.js
@@ -1,12 +1,22 @@
-import { generateWhereClause } from "../utils";
+import { generateWhereClause } from '../utils';
+import generateUpdate from './update';
 
 const generateDelete = (model, options) => {
-  const { where } = options;
+    const { where } = options;
 
-  const whereClause = generateWhereClause(model.modelName, where);
-  const deleteQuery = `DELETE FROM ${model.modelName}${whereClause};`;
+    if (model.paranoid) {
+        const updateQuery = generateUpdate(
+            model,
+            { deletedAt: new Date() },
+            options
+        );
+        return updateQuery;
+    }
 
-  return deleteQuery;
+    const whereClause = generateWhereClause(model.modelName, where);
+    const deleteQuery = `DELETE FROM ${model.modelName}${whereClause};`;
+
+    return deleteQuery;
 };
 
 export default generateDelete;
